# backend-master/apiDocumentation.yaml
swagger: "2.0"
info:
  description: "Booking-mini description"
  version: "1.0-SNAPSHOT"
  title: "Booking-mini"
host: "localhost"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
...
  /hotel:
    post:
      tags:
      - "hotel"
      summary: "Add new hotel"
      description: "This can only be done by the landlord "
      operationId: "createHotel"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Info about hotel"
        required: true
        schema:
          $ref: '#/definitions/NewHotel'
      responses:
        201:
          description: "Successful hotel creation"
          schema:
            $ref: '#/definitions/Hotel'
        400:
          description: "Data is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
        401:
          description:   "Unauthorized"
        403:
          description:  "Forbidden"
          
    get:
      tags:
      - "hotel"
      summary: "Get all hotels"
      operationId: "getAllHotels"
      produces:
      - "application/json"
      responses:
        200:
          description: "Hotels found"
          schema:
              type: "object"
              properties:
                amount:
                  type: "integer"
                list:
                  $ref: '#/definitions/Hotel'
        400:
          description: "Data is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
  /hotel/{id}:
    get:
      tags:
      - "hotel"
      summary: "Get hotel by id"
      operationId: "getHotelById"
      produces:
      - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Hotel's id. "
          required: true
          type: "integer"
      responses:
        200:
          description: "Hotel found"
          schema:
              $ref: '#/definitions/DetailedHotel'
        400:
          description: "Data is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
      - "hotel"
      summary: "Delete hotel"
      description: "This can only be done by hotel creator."
      operationId: "deleteHotel"
      parameters:
      - name: "id"
        in: "path"
        description: "Hotel's id. "
        required: true
        type: "integer"
      responses:
        "200":
          description: "Succesfull deletion"
        "400":
          description: "Invalid id supplied"
        "403":
          description: "Forbidden"
        "404":
          description: "Hotel not found"
    put:
      tags:
      - "hotel"
      summary: "Update hotel info"
      description: "This can only be done by hotel creator."
      operationId: "updateHotel"
      parameters:
      - name: "id"
        in: "path"
        description: "User's id. "
        required: true
        type: "integer"
      - name : "updatedHotel"  
        in: "body"
        required : true
        schema:
         $ref: '#/definitions/NewHotel'
      responses:
        "200":
          description: "Succesfull update"
          schema:
              type: "integer"
              description: "hotel id"
        "400":
          description: "Invalid id supplied"
        "403":
          description: "Forbidden"  
        "404":
          description: "Hotel not found"   
  /hotel/{country}/{city}:
    get:
      tags:
      - "hotel"
      summary: "Get hotels by location"
      operationId: "getHotelByLocation"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "country"
        required: true
        type: "string"
      - in: "path"
        name: "city"
        required: true
        type: "string"
      responses:
        200:
          description: "Hotels found"
          schema:
              type: "object"
              properties:
                amount:
                  type: "integer"
                list:
                  $ref: '#/definitions/Hotel'
        400:
          description: "Data is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
  /hotel/{country}/{city}/byDate:
    get:
      tags:
      - "hotel"
      summary: "Get hotel info by location and dates"
      operationId: "getHotelByDateAndLocation"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "country"
        required: true
        type: "string"
      - in: "path"
        name: "city"
        required: true
        type: "string"
      - in: "body"
        name: "dates"
        schema:
            type: "object"
            properties:
              arrivalDate:
                type: "string"
                format: "date"
              departureDate:
                type: "string"
                format: "date" 
     
               
      responses:
        200:
          description: "Hotels found"
          schema:
              type: "object"
              properties:
                amount:
                  type: "integer"
                list:
                  $ref: '#/definitions/Hotel'
        400:
          description: "Data is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
  /hotel/createdby/{username}:
    get:
      tags:
      - "hotel"
      summary: "Get hotels created by particular user."
      operationId: "getHotelbyUser"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "username"
        required: true
        type: "string"
      responses:
        200:
          description: "Hotels found"
          schema:
              type: "object"
              properties:
                amount:
                  type: "integer"
                list:
                  $ref: '#/definitions/Hotel'
        400:
          description: "Data is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
  /hotels/{id}:
    get:
      tags:
      - "hotel"
      summary: "Get hotel by id"
      description: ""
      operationId: "getHotel"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Hotel found"
          schema:
           $ref: '#/definitions/Hotel'
        400:
          description: "Data is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
...
  /comments/hotel/{hotelid}:
    get:
      tags:
      - "comments"
      summary: "Get comments for particular hotel"
      operationId: "commentsByHotel"
      parameters:
      - in: "path"
        name: "hotelid"
        type: "integer"
        description: "Hotel for which comments you want "
        required: true
      responses:
        200:
          description: "Comments found"
          schema:
              type: "object"
              properties:
                amount:
                  type: "integer"
                list:
                  $ref: '#/definitions/Comment'
        400:
          description: "Data is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
        404:
          description: "Hotel not found"        
...
  /image/{hotelId}:
    get:
      tags:
      - "image"
      summary: "Get images for hotel"
      operationId: "getImages"
      parameters:
      - in: "path"
        name: "hotelId"
        type: "integer"
        description: "Hotel's id "
        required: true
      responses:
        201:
          description: "Images found"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/Image' 
        400:
          description: "hotel id is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
        404:
          description:   "Hotel not found"    
    put:
      tags:
      - "image"
      summary: "Add images for hotel"
      description: "This can only be done by hotel's creator"
      operationId: "addImages"
      parameters:
      - in: "path"
        name: "hotelId"
        type: "integer"
        description: "Hotel's id "
        required: true
      - in: "body" 
        name: "files"
        schema:
          type: "array"
          items: 
           $ref: '#/definitions/Image'
      responses:
        201:
          description: "Images added"

        400:
          description: "Files are invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
        404:
          description:   "Hotel not found" 
        403:
          description: "Forbidden"
  /image/first/{hotelId}:
    get:
      tags:
      - "image"
      summary: "Get icon for hotel in list"
      operationId: "getFirstImage"
      parameters:
      - in: "path"
        name: "hotelId"
        type: "integer"
        description: "Hotel's id "
        required: true
      responses:
        200:
          description: "Image found"
          schema:
              $ref: '#/definitions/Image' 
        400:
          description: "hotel id is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
        404:
          description:  "Hotel not found" 
  /chat:
    post:
      tags:
      - "chat"
      summary: "Send message"
      operationId: "sendMessage"
      parameters:
      - in: "body"
        name: "message"
        schema:
          $ref: '#/definitions/Message'
      responses:
        200:
          description: "Message sent"
        400:
          description: "Message is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
        404:
            description:  "Chat not found"        
  /chat/create:
    post:
      tags:
      - "chat"
      summary: "Start chat"
      operationId: "createChat"
      parameters:
      - in: "body"
        name: "chatModel"
        schema:
          $ref: '#/definitions/ChatModel'
      responses:
        201:
          description: "Chat created"
          schema:
              $ref: '#/definitions/ChatModel' 
        400:
          description: "One of user's id is invalid"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ErrorResponse'
        404:
          description:  "User not found" 
  /chat/mychats:
    get:
      tags:
      - "chat"
      summary: "Get list of my chats"
      operationId: "getMyChats"
      responses:
        200:
          description: "Your chats"
          schema:
            type: array
            items:
              $ref: '#/definitions/ChatListItem'
        403:
          description:  "Not authenticated" 
  /chat/messages/{chatId}:
    get:
      tags:
      - "chat"
      summary: "Get messages for chat"
      operationId: "getChatMessages"
      parameters:
      - in: "path"
        name: "chatId"
        type: "integer"
        description: "Chat's id "
        required: true
      responses:
        200:
          description: "Messages found"
          schema:
            type: array
            items:
              $ref: '#/definitions/Message'
        403:
          description:  "You dont have access to that chat"        
        404:  
          description:  "Chat not found"        
definitions:
  Hotel:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        minimum: 1
      name:
        type: "string"
        maxLength:  30
        minLength: 4
        example: "Zastava"
      location:
        type: "object"
        properties:
          country:
            type: "string"
            minLength: 3
            example: "Russia"
          city:
            type: "string"
            minLength: 3
            example: "Voronezh"
          street:
            type: "string"
            example: "Moskovsky prospect"
          house:
            type: "string"
            example: "21d"
            maxLength: 10
      description:
        type: "string"
        minLength: 30
        maxLength: 500
        example: "Zastava Hotel is located in Voronezh, 2.1 km from Central City Stadium and 3.3 km from Sunny Paradise Shopping Center. Offering city views, air-conditioned rooms have a desk and free Wi-Fi."
      amenities:
        type: array
        items:
          type: string
        example: ["free wifi","parking", "near beach"]
      createdAt:
        type: "string"
        format: "date-time"
        example: "2020-07-13T10:30:00Z"
  DetailedHotel:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        minimum: 1
      name:
        type: "string"
        maxLength:  30
        minLength: 4
        example: "Zastava"
      location:
        type: "object"
        properties:
          country:
            type: "string"
            minLength: 3
            example: "Russia"
          city:
            type: "string"
            minLength: 3
            example: "Voronezh"
          street:
            type: "string"
            example: "Moskovsky prospect"
          house:
            type: "string"
            example: "21d"
            maxLength: 10
      description:
        type: "string"
        minLength: 30
        maxLength: 500
        example: "Zastava Hotel is located in Voronezh, 2.1 km from Central City Stadium and 3.3 km from Sunny Paradise Shopping Center. Offering city views, air-conditioned rooms have a desk and free Wi-Fi."
      amenities:
        type: array
        items:
          type: string
        example: ["free wifi","parking", "near beach"]
      createdAt:
        type: "string"
        format: "date-time"
        example: "2020-07-13T10:30:00Z" 
      CreatorName:
        type: "string"
        maxLength:  30
        minLength: 4
        example: "Vasya"   
      CreatorId:
        type: "integer"
        minimum: 1
        example: 1 
  NewHotel:
    type: "object"
    properties:
      name:
        type: "string"
        maxLength:  30
        minLength: 4
        example: "Zastava"
      location:
        type: "object"
        properties:
          country:
            type: "string"
            minLength: 3
            example: "Russia"
          city:
            type: "string"
            minLength: 3
            example: "Voronezh"
          street:
            type: "string"
            example: "Moskovsky prospect"
          house:
            type: "string"
            example: "21d"
            maxLength: 10
      description:
        type: "string"
        minLength: 30
        maxLength: 500
        example: "Zastava Hotel is located in Voronezh, 2.1 km from Central City Stadium and 3.3 km from Sunny Paradise Shopping Center. Offering city views, air-conditioned rooms have a desk and free Wi-Fi."
      amenities:
        type: array
        items:
          type: string
        example: ["free wifi","parking", "near beach"]
...
  Comment:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      hotelId:
        type: "integer"
        format: "int64"
        minimum: 1
      createdBy:
        type: "string"
        minLength: 5
        maxLength: 64
        example: "user1"
      title:
        type: "string"
        minLength: 1
        maxLength: 50
      text:
        type: "string"
        minLength: 10
        maxLength: 500
      createdAt:
        type: "string"
        format: "date-time"
        example: "2020-07-13T10:30:00Z"
      rating:
        type: "integer"
        minimum: 1
        maximum: 5
...
  Image:
    type: "object"
    properties:
      lg: 
       type: "string"
       description: "Path to image in large resolution"
       example: "hotel-image1.jpg"
  ChatListItem: 
    type: "object"
    properties:
      chatId: 
       type: "integer"
       description: "Chat's id"
       example: 32
      partnerName: 
       type: "string"
       description: "Name of your chat partner"
       example: "sampleUser2"
      lastMessage: 
       type: "string"
       description: "Last message on chat"
       example: "Hello there"
  ChatModel: 
    type: "object"
    properties:
      chatId: 
       type: "integer"
       description: "Chat's id"
       example: 32
      landlordId: 
       type: "integer"
       description: "Landlord's id"
       example: 3
  Message: 
    type: "object"
    properties:
      chatId: 
       type: "integer"
       description: "Chat's id"
       example: 32
      sentBy: 
       type: "string"
       description: "User who sent the message"
       example: "sampleUser2"
      message: 
       type: "string"
       description: "Message text"
       example: "Hello there "
      sentAt: 
       type: "string"
       format: date
       description: "When the message was sent"
  ErrorResponse:
    type: "object"
    properties:
      code:
        type: "string"
      description:
        type: "string"
      payload:
        type: "object"